% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns.R
\name{cov_mat}
\alias{cov_mat}
\title{The rank-based estimate of the correlation matrix}
\usage{
cov_mat(dat, ordinal_index)
}
\arguments{
\item{dat}{a numeric matrix or data frame, of dimensions n by p. If matrix, each row is a length-p vector representing one observation on p variables.}

\item{ordinal_index}{a numeric or logical vector. The indices of columns in \code{dat} that are of ordinal values (rather than continuous).}
}
\description{
\code{cov_mat} computes the estimate of the correlation matrix between the columns of \code{dat}.
}
\examples{
dat <- cbind(c(0,1,0,0,1), c(1,0,1,1,0), rnorm(5), rnorm(5))
cov_mat(dat, c(1,2))
}
