% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns.R
\name{d_hat_vec}
\alias{d_hat_vec}
\title{Moment estimate of the cutoffs for discretization.}
\usage{
d_hat_vec(x)
}
\arguments{
\item{x}{a numeric vector. Observed values of a variable. Must be of (ordinal) discrete values.}
}
\value{
The moment estimate of the cutoffs.
}
\description{
Moment estimate of the cutoffs for discretization.
}
\examples{
x <- c(0, 1, 0, 0, 1)
d_hat_vec(x)
}
