% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns.R
\name{G_ordinal}
\alias{G_ordinal}
\title{Expectation of sample Kendall's tau.}
\usage{
G_ordinal(r, dj1, dj2, dk1, dk2)
}
\arguments{
\item{r}{Correlation coefficient between the two latent variables.}

\item{dj1}{The cutoff \eqn{\Delta_j^1}.}

\item{dj2}{The cutoff \eqn{\Delta_j^2}.}

\item{dk1}{The cutoff \eqn{\Delta_k^1}.}

\item{dk2}{The cutoff \eqn{\Delta_k^2}.}
}
\value{
The population version of Kendall's tau.
}
\description{
Expectation of sample Kendall's tau.
}
\examples{
G_ordinal(.5, -0.3, 0.3, 0, Inf)
}
