% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns.R
\name{predict}
\alias{predict}
\title{Predict y for given x using GC-Scout method.}
\usage{
predict(x_new, x, y, betahat)
}
\arguments{
\item{x_new}{A numeric matrix. Must have the same number of columns as \code{x}}

\item{x}{A numeric matrix. Used for training.}

\item{y}{A numeric vector. Used for training.}

\item{betahat}{A numeric matrix or vector. If vector, its length must be equal to \code{x}'s number of columns. If matrix, each column represents one beta.}
}
\value{
Given the observed pairs (\code{x},\code{y}), and the
\eqn{\beta} estimate \code{betahat}, one can predict new response
values for new covariates \code{x_new}, under latent Gaussian Copula model.
Specifically, this function implements Algorithm 2 from Quan et al (2019).
}
\description{
Predict y for given x using GC-Scout method.
}
\examples{
n <- 200; d <- ncol(Sigma_witten); d3 <- 4; delta = c(qnorm(0.5), qnorm(0.75))
dat <- mvtnorm::rmvnorm(n, sigma = Sigma_witten)
y_tilde = dat[,d]
y = y_tilde; y = exp(y_tilde);
x_tilde = dat[,-d]; x = x_tilde^5
x[,1:d3] = as.numeric(cut(x[,1:d3], breaks = c(-Inf,(delta)^5,Inf)))-1
x_train = x[1:40,]; y_train = y[1:40]
x_test = x[41:n,]; y_test = y[41:n]
predict(x_test, x_train, y_train, c(0.5,0.3,0,0,0,0,-0.2,-0.2))
}
