% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns.R
\name{tau_hat}
\alias{tau_hat}
\title{Kendall's tau-a.}
\usage{
tau_hat(x, y, j, k, n)
}
\arguments{
\item{x}{A numeric matrix.}

\item{y}{A numeric matrix.}

\item{j}{an integer. Indicating which column of \code{x} is to be considered for this function.}

\item{k}{an integer. Indicating which column of \code{y} is to be considered.}

\item{n}{an integer. The number of observations. Has to be same for \code{x} and \code{y}.}
}
\value{
The Kendall's tau-a correlation coefficient between the \code{j}-th column of \code{x} and the \code{k}-th column of \code{y}.
}
\description{
Compute the Kendall rank correlation coefficient (Tau-a statistic) between two variables.
}
\examples{
x <- cbind(c(0,1,0,0,1), c(1,1,0,1,0))
y <- cbind(rnorm(5), rnorm(5))
tau_hat(x, x, 1, 2, 5)
tau_hat(x, y, 1, 1, 5)
}
